edited deliveries_controller.rb

+++ app/controllers/deliveries_controller.rb
@@ -1,17 +1,20 @@
 class DeliveriesController < ApplicationController
+  # before_action :authenticate_user!
   def index
-    # before_action :authenticate_user!
+    
       # if there is no user redirect to the login page
-      if current_user == nil
-        redirect_to("/users/sign_in")
-      else
+       if current_user == nil
+         redirect_to("/users/sign_in")
+       return
+       end
     matching_deliveries = Delivery.all
 
-    @list_of_deliveries = matching_deliveries.order({ :created_at => :desc })
+     @list_of_deliveries = matching_deliveries.order({ :created_at => :desc })
+  
 
     render({ :template => "deliveries/index" })
+   
   end
-end
 
   def show
     the_id = params.fetch("path_id")
@@ -24,18 +27,18 @@ end
   end
 
   def create
-    the_delivery = Delivery.new
-    the_delivery.user_id = current_user.id
-    the_delivery.description = params.fetch("query_description")
-    the_delivery.supposed_to_arrive_on = params.fetch("query_supposed_to_arrive_on")
-    the_delivery.details = params.fetch("query_details")
-    the_delivery.arrived = false
+    delivery = Delivery.new
+    delivery.user_id = current_user.id
+    delivery.description = params.fetch("query_description")
+    delivery.supposed_to_arrive_on = params.fetch("query_supposed_to_arrive_on")
+    delivery.details = params.fetch("query_details")
+    delivery.arrived = false
 
-    if the_delivery.valid?
-      the_delivery.save
+    if delivery.valid?
+      delivery.save
       redirect_to("/deliveries", { :notice => "Delivery created successfully." })
     else
-      redirect_to("/deliveries", { :alert => the_delivery.errors.full_messages.to_sentence })
+      redirect_to("/deliveries", { :alert => delivery.errors.full_messages.to_sentence })
     end
   end
 
@@ -46,13 +49,13 @@ end
     # the_delivery.description = params.fetch("query_description")
     # the_delivery.supposed_to_arrive_on = params.fetch("query_supposed_to_arrive_on")
     # the_delivery.details = params.fetch("query_details")
-    # the_delivery.arrived = params.fetch("query_arrived")
+     the_delivery.arrived = params.fetch("query_arrived")
 
     if the_delivery.valid?
       the_delivery.save
-      redirect_to("/deliveries/#{the_delivery.id}", { :notice => "Delivery updated successfully."} )
+      redirect_to("/", { :notice => "Delivery updated successfully."} )
     else
-      redirect_to("/deliveries/#{the_delivery.id}", { :alert => the_delivery.errors.full_messages.to_sentence })
+      redirect_to("/", { :alert => the_delivery.errors.full_messages.to_sentence })
     end
   end
 
@@ -60,8 +63,16 @@ end
     the_id = params.fetch("path_id")
     the_delivery = Delivery.where({ :id => the_id }).at(0)
 
-    the_delivery.destroy
 
-    redirect_to("/deliveries", { :notice => "Delivery deleted successfully."} )
+  if the_delivery
+    the_delivery.destroy
+    redirect_to deliveries_path, notice: "Delivery deleted successfully."
+  else
+    redirect_to deliveries_path, alert: "Delivery not found."
+  end
+
+    # the_delivery.destroy
+
+    # redirect_to("/deliveries", { :notice => "Delivery deleted successfully."} )
   end
 end
